backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi
uvicorn[standard]
sentence-transformers
pymupdf
python-multipart
faiss-cpu
requests
beautifulsoup4
python-docx
google-generativeai==0.5.4
torch==2.7.1+cpu
python-dotenv
passlib[bcrypt]
pyjwt
